* commit '60f76980aab92f0e0730170b99a2363acd3c97fa': (108 commits)
  Update version numbers and remove empty sections in RELEASE.txt.
  Restore line 1154: C2Cppfunction_map.mht to HTML_EXTRA_FILES.
  Update version for 1.8.20 release.
  Fixed documentation warnings
  HDFFV-10329 Correct LINK INTERFACE and VS20013 compile
  Switch default build mode from development to production. Increment version string to pre2.
  Add RELEASE.txt note for HDFFV-10274. (cherry picked from commit f7a7d0a00613cba997212fa8032091629a678797)
  Initialize hid_t variables in test_Attr_bug9. (cherry picked from commit 97bc393449f3fe02c5992872ab2842e12f611ef0)
  Fix HDFFV-10274.  When deleting all (or almost all) of the messages in an object header chunk, where the total amount deleted was greater than 64K, an error would occur due to an off by one error in the code that handled that case.  Fixed this and added a test case. (cherry picked from commit 1b2c2ca9a6a7d7e1fcd5c3302e203f2e2dabf0af)
  Fix linking of libs
  Changed file path to relative.
  iAdd NAG to supported compilers in RELEASE.txt.
  Add description to RELEASE.txt for HDFFV-10323
  Correct issue number
  Set version to 1.8.20-pre1
  Update supported and tested platforms in RELEASE.txt.
  HDFFV-10321 Merge from 1.8
  Incremented lib file .so numbers: revision for all except c++ for code change; current incremented, revision and age set to 0 for c++ due to symbols added and removed.
  updated issues related to Fortran
  Incremented lib file .so numbers: revision for all except c++ for dcode change; current incremented, revision and age set to 0 for c++ due to symbols added and removed.
  ...

+++ tools/lib/io_timer.c
@@ -22,9 +22,6 @@
  * This is a module of useful timing functions for performance testing.
  */
 
-#include <stdio.h>
-#include <stdlib.h>
-
 #include "H5private.h"
 #include "hdf5.h"
 
@@ -53,9 +50,9 @@ io_time_t   *timer_g;            /* timer: global for stub functions     */
 static double sub_time(struct timeval* a, struct timeval* b)
 {
     return (((double)a->tv_sec +
-     ((double)a->tv_usec) / MICROSECOND) -
+     ((double)a->tv_usec) / (double)MICROSECOND) -
   ((double)b->tv_sec +
-   ((double)b->tv_usec) / MICROSECOND));
+   ((double)b->tv_usec) / (double)MICROSECOND));
 }
 
 
@@ -195,8 +192,9 @@ set_time(io_time_t *pt, timer_type t, int start_stop)
 
             }
 	break;
+
     default:
-	HDfprintf(stderr, "Unknown time clock type (%d)\n", pt->type);
+	    HDfprintf(stderr, "Unknown time clock type (%d)\n", pt->type);
 	    return NULL;
     } /* end switch */
